+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Exceptions are acceptable depending on the circumstances (critical bug fixes tha
 
 ## [Unreleased]
 
+### Changed
+
+- corrected Azure DevOps cloning strategy to avoid errors while using tokens
+
 ## [2.15.0] - 2025-12-16
 
 ### Added